+++ CHANGELOG.md
@@ -2,10 +2,9 @@
 
 - Add `pressOn` namespace which includes `Character`, `Space`, `Delete`, `Backspace`, `Enter`, `Tab`, `ArrowLeft`, `ArrowRight`, `ArrowUp`, `ArrowDown`, `Undo`, `Redo`, `SelectAll` methods. Using these methods you can imitate physical key presses.
 - Add `typeText`: type text at current caret position.
-- Add `isTextSelected`: indicates that text is selected (at least on one line).
+- Add `isTextSelected`: indicates what text is selected (at least on one line).
 - Add `deleteSelection`: remove current selection.
 - Add type definitions.
-- Add `ESlint`.
 
 
 # 2.0.0 (December 24, 2020)