+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 2.1.0
+# 2.1.0 (December 28, 2020)
 
 - Add `pressOn` namespace which includes `Character`, `Space`, `Delete`, `Backspace`, `Enter`, `Tab`, `ArrowLeft`, `ArrowRight`, `ArrowUp`, `ArrowDown`, `Undo`, `Redo`, `SelectAll` methods. Using these methods you can imitate physical key presses.
 - Add `typeText`: type text at current caret position.